Refactor tests and build lib separation
Introduce `repo_structure_diff_scan.py` for incremental diff scanning, enhancing directory structure validation. Refactor and consolidate exceptions and data classes into `repo_structure_lib.py` for cleaner code organization and maintenance.
